requirements.txt
requests
pandas
pymupdf
tqdm
openpyxl
python-dotenv
groq
streamlit